Refactor: dataset DTO renaming and restructuring
