--- src/core/domain/useCases/validators/NewResourceValidator.ts
@@ -1,4 +1,0 @@
-export interface NewResourceValidator {
-  // eslint-disable-next-line  @typescript-eslint/no-explicit-any
-  validate(...args: any[]): void
-}
