+++ src/core/domain/useCases/validators/ResourceValidator.ts
@@ -1,0 +1,3 @@
+export interface ResourceValidator {
+  validate(...args: unknown[]): void
+}
