+++ src/datasets/domain/useCases/CreateDataset.ts
@@ -1,6 +1,6 @@
 import { IDatasetsRepository } from '../repositories/IDatasetsRepository'
-import { NewDatasetDTO } from '../dtos/NewDatasetDTO'
-import { NewResourceValidator } from '../../../core/domain/useCases/validators/NewResourceValidator'
+import { DatasetDTO } from '../dtos/DatasetDTO'
+import { ResourceValidator } from '../../../core/domain/useCases/validators/ResourceValidator'
 import { IMetadataBlocksRepository } from '../../../metadataBlocks/domain/repositories/IMetadataBlocksRepository'
 import { CreatedDatasetIdentifiers } from '../models/CreatedDatasetIdentifiers'
 import { ROOT_COLLECTION_ALIAS } from '../../../collections/domain/models/Collection'
@@ -10,7 +10,7 @@ export class CreateDataset extends DatasetWriteUseCase<CreatedDatasetIdentifiers
   constructor(
     datasetsRepository: IDatasetsRepository,
     metadataBlocksRepository: IMetadataBlocksRepository,
-    newDatasetValidator: NewResourceValidator
+    newDatasetValidator: ResourceValidator
   ) {
     super(datasetsRepository, metadataBlocksRepository, newDatasetValidator)
   }
@@ -18,7 +18,7 @@ export class CreateDataset extends DatasetWriteUseCase<CreatedDatasetIdentifiers
   /**
    * Creates a new Dataset in a collection, given a NewDatasetDTO object and an optional collection identifier, which defaults to root.
    *
-   * @param {NewDatasetDTO} [newDataset] - NewDatasetDTO object including the new dataset metadata field values for each metadata block.
+   * @param {DatasetDTO} [newDataset] - DatasetDTO object including the new dataset metadata field values for each metadata block.
    * @param {string} [collectionId] - Specifies the collection identifier where the new dataset should be created (optional, defaults to root).
    * @returns {Promise<CreatedDatasetIdentifiers>}
    * @throws {ResourceValidationError} - If there are validation errors related to the provided information.
@@ -26,7 +26,7 @@ export class CreateDataset extends DatasetWriteUseCase<CreatedDatasetIdentifiers
    * @throws {WriteError} - If there are errors while writing data.
    */
   async execute(
-    newDataset: NewDatasetDTO,
+    newDataset: DatasetDTO,
     collectionId = ROOT_COLLECTION_ALIAS
   ): Promise<CreatedDatasetIdentifiers> {
     const metadataBlocks = await this.getNewDatasetMetadataBlocks(newDataset)
