+++ test/testHelpers/TestConstants.ts
@@ -1,4 +1,4 @@
-import { NewDatasetDTO } from '../../src/datasets/domain/dtos/NewDatasetDTO'
+import { DatasetDTO } from '../../src/datasets/domain/dtos/DatasetDTO'
 
 export class TestConstants {
   static readonly TEST_API_URL = 'http://localhost:8080/api/v1'
@@ -49,7 +49,7 @@ export class TestConstants {
   static readonly TEST_DUMMY_COLLECTION_ALIAS = 'dummyCollectionId'
   static readonly TEST_CREATED_COLLECTION_ALIAS_1 = 'testCollection1'
   static readonly TEST_CREATED_COLLECTION_ALIAS_2 = 'testCollection2'
-  static readonly TEST_NEW_DATASET_DTO: NewDatasetDTO = {
+  static readonly TEST_NEW_DATASET_DTO: DatasetDTO = {
     license: {
       name: 'CC0 1.0',
       uri: 'http://creativecommons.org/publicdomain/zero/1.0',
